refactor(merge): rolling.go 按域拆分 — 1897 行 → 5 个按域文件 (#611)
纯移动无逻辑变化：rolling.go(618, 协调器/配置/生命周期/磁盘工具) + rolling_backfill.go(384, 历史回填引擎+单例清理) + rolling_batch.go(520, MP4/音频批量合并+兼容性分组) + rolling_live.go(482, 事件循环+实时段合并) + rolling_bucket.go(201, bucket 创建/追加 IO)
